I haven't committed in a while
There are now a set of files which are independent of the input and output formats
of the TensorFlow model: collect_initial_data.py, check_data.py, prepare_data.py,
train_new_model.py, run_model.py and collect_correction_data.py.

Almost all of the model-specific code is now stored in modules named after the
model format. E.g. multihot_3.py and
multihot_4_discarding_forwards_and_straight_ons.py. In addition to the input and
output shapes of the model, these files also define the format of the training
data and what will be discarded during data collection, amongst other things.

I think I might effectively be using these modules as classes. I'll look into using
classes to store model-specific code. That may be better for state information
which I want to be present when the model is running (e.g. a rolling average of
the model output).

+++ check_data.py
@@ -1,29 +1,35 @@
 import numpy as np
 import cv2
 from time import time, sleep
+import os
 
-from common import INITIAL_DATA_FILE_NAME, DISPLAY_WIDTH, DISPLAY_HEIGHT, OUTPUT_LENGTH
+from common import INITIAL_DATA_FILE_NAME, DISPLAY_WIDTH, DISPLAY_HEIGHT, CORRECTION_DATA_FILE_NAME
 
-initial_data = np.load(INITIAL_DATA_FILE_NAME)
+from multihot_3 import output_row_to_wasd_string
 
-for frame in initial_data:
+data = None  # Stops "'data' can be undefined"
+if os.path.isfile(CORRECTION_DATA_FILE_NAME):
+    choice = input('Check correction data? (y/n)\n')
+    if choice == 'y':
+        print('Loading correction data')
+        data = np.load(CORRECTION_DATA_FILE_NAME)
+    else:
+        print('Loading initial data')
+        data = np.load(INITIAL_DATA_FILE_NAME)
+        print(data.shape)
+        exit()
+
+for datum in data:
     start_time = time()
-    cv2.imshow('initial data', cv2.resize(frame[:-1], (DISPLAY_WIDTH, DISPLAY_HEIGHT), interpolation=cv2.INTER_NEAREST))
+    frame = datum[:-1]
+    output_row = datum[-1]
+    cv2.imshow('data', cv2.resize(frame, (DISPLAY_WIDTH, DISPLAY_HEIGHT), interpolation=cv2.INTER_NEAREST))
     if cv2.waitKey(25) & 0xFF == ord('q'):
         cv2.destroyAllWindows()
         break
 
-    outputs = list(frame[-1].reshape((-1, 4)))
-    for i in range(len(outputs)):
-        a = np.argmax(outputs[i])
-        if a == 0:
-            outputs[i] = 'w'
-        elif a == 1:
-            outputs[i] = 'wa'
-        elif a == 2:
-            outputs[i] = 'wd'
-        elif a == 3:
-            outputs[i] = 's'
+    print(output_row_to_wasd_string(output_row))
 
-    print(outputs)
-    sleep(max(0., 1/5 - (time() - start_time)))
+    # sleep(max(0., 1/5 - (time() - start_time)))
+
+cv2.destroyAllWindows()
